create-get-similar-products-service

+++ src/main/java/com/tmartinez/similarproducts/domain/model/Product.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 public class Product {
     private String id;
     private String name;
-    private Float price;
+    private Double price;
     private Boolean availability;
 }
